+++ .fttemplates/testcomp2/[FFSName].js
@@ -1,5 +1,0 @@
-const <FFSName> = () => {
-  return <div />
-}
-
-export default <FFSName>