+++ examples/TS Function Component/[FTName].ts
@@ -1,0 +1,10 @@
+import React from 'react';
+import './[FTName].css';
+
+type <FTName | capitalize>Props = {}
+
+const <FTName | capitalize> = (props: <FTName | capitalize>Props) => {
+  return <div></div>;
+};
+
+export default <FTName | capitalize>;