+++ examples/TS Function Component/index.ts
@@ -1,0 +1,3 @@
+import <FTName | capitalize> from './[FTName]';
+
+export default <FTName | capitalize>;