+++ examples/Test File/[FTName % capitalize].test.js
@@ -1,0 +1,7 @@
+import <FTName | capitalize> from '../<containingFolder>/<FTName>'
+
+describe('<FTName | capitalize>', () => {
+  it('should', () => {
+
+  });
+});